+++ gupb/controller/BenadrylowyBarabasz/__init__.py
@@ -1,0 +1,10 @@
+from .barabasz import BarabaszController
+
+__all__ = [
+    'BarabaszController',
+    'POTENTIAL_CONTROLLERS'
+]
+
+POTENTIAL_CONTROLLERS = [
+    BarabaszController("BenadrylowyBarabasz"),
+]
